requirements.txt
aioredis==1.2.0
amqp==2.4.2
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
billiard==3.6.0.0
cairocffi==1.0.2
CairoSVG==2.3.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
Cython==0.29.10
daphne==2.2.5
decorator==4.3.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.7
django-common-helpers==0.9.2
django-cron==0.5.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-rest-swagger==2.2.0
django-stubs==0.12.1
djangorestframework==3.9.1
djangorestframework-stubs==0.3.0
et-xmlfile==1.0.1
gunicorn==19.9.0
hiredis==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipython==7.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.13.2
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1
msgpack==0.6.1
mypy==0.711
mypy-extensions==0.4.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.8
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
PyHamcrest==1.9.0
pyminifier==2.1
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2018.9
PyYAML==5.1
raven==6.10.0
redis==3.2.1
reportlab==3.5.19
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sorl-thumbnail==12.5.0
soupsieve==1.9.1
tablib==0.13.0
tinycss2==1.0.2
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.4.0
typing-extensions==3.7.4
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
WeasyPrint==47
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
